Fixed click-to-move, weird scrolling behavior, cleanup code
This is in a more or less usable state now and doesn't crash
Textadept in edge cases anymore, but some weird interaction with
view:goto_buffer is still causing things to desync and behave oddly.

Updated the bugs section of the README and added instructions to
click on the minimap to move the view around.