Changed rectangular select calculation method
Changed the way that the select rectangle is calculated to use a
Scintilla feature to find the last column instead of iterating,
should be a lot faster. I have a better way of doing this now but
I wanted to commit this in case I want to revisit this style later.