Fixed some crashing bugs, made cleanup better. Updated README.
Tried implementing entire line selection for the minimap highlight
but it broke the click to jump function and I couldn't figure out
how to make it work after playing with it for a few hours so I'm
just going to leave it as is. Performance doesn't seem to be much worse.

Minimap now cleans up after itself better on a reset and when unsplitting.
Still doesn't work with session saving though.